refactor(au_qld_ecq_elections): rename the GCP dataset id to carry the jurisdiction
The Australian state election datasets now follow au_<state>_<agency>_elections,
so the Queensland GCP dataset id becomes au_qld_ecq_elections. Renames the models
and pipelines package directories, the 11 dbt models, the dbt_project.yml entry,
every schema/alias config and set_datalake_project staging reference, and the
scratch data root default.

The backend dataset slug (qld_elections) and the organization (au_ecq) are
unchanged; only the GCP dataset id moves.
