+++ models/au_qld_ecq_elections/code/architecture/candidate.csv
