+++ models/au_qld_ecq_elections/code/build_dbt.py
@@ -1,10 +1,10 @@
-"""Generate the au_ecq_elections dbt models and schema.yml.
+"""Generate the au_qld_ecq_elections dbt models and schema.yml.
 
-Both are derived from ``pipelines/datasets/au_ecq_elections/schema.py`` so the
+Both are derived from ``pipelines/datasets/au_qld_ecq_elections/schema.py`` so the
 architecture, the cleaning transform, the staging schema and the models cannot drift
 apart. Regenerate rather than hand-editing the SQL.
 
-Run:  PYTHONPATH=. ~/.venvs/bd-pipelines/bin/python models/au_ecq_elections/code/build_dbt.py
+Run:  PYTHONPATH=. ~/.venvs/bd-pipelines/bin/python models/au_qld_ecq_elections/code/build_dbt.py
 
 The committed files are the post-hook form: ``sqlfmt`` and ``yamlfix`` rewrite this
 script's output on commit (list flow style, line wrapping). Run pre-commit on the
@@ -15,8 +15,11 @@ from __future__ import annotations
 
 import textwrap
 
-from pipelines.datasets.au_ecq_elections import schema
-from pipelines.datasets.au_ecq_elections.constants import REPO_ROOT, constants
+from pipelines.datasets.au_qld_ecq_elections import schema
+from pipelines.datasets.au_qld_ecq_elections.constants import (
+    REPO_ROOT,
+    constants,
+)
 
 DATASET = constants.DATASET_ID.value
 MODEL_DIR = REPO_ROOT / "models" / DATASET
